add AppleBCM5701Ethernet.kext from 20A4300b
kral84 on the MacRumors forum is reporting that doing a speed test
causes his 2011 iMac's Ethernet connection to stop working until he
unplugs and replugs it. I'm unable to reproduce this on my 2011
MacBook Pro, but it's possible that using the kext from Catalina 10.15.6
instead of Big Sur dev beta 2 is causing this.

(Another commit will be needed to change patch-kexts.sh over to this kext.)